move snacks Queries into 'snacks/' domain
