+++ src/resolvers/Query/index.ts
@@ -3,7 +3,7 @@ export { getPage } from "./getPage"
 export { listTables } from "./listTables"
 export { queryMovies } from "./queryMovies"
 export { queryPages } from "./queryPages"
-export { querySnacks } from "./querySnacks"
+export { querySnacks } from "./snacks/querySnacks"
 export { scanIpsTable } from "./scanIpsTable"
 export { scanPagesTable } from "./scanPagesTable"
-export { scanSnacksTable } from "./scanSnacksTable"
+export { scanSnacksTable } from "./snacks/scanSnacksTable"
