+++ src/resolvers/Query/snacks/scanSnacksTable.ts
